same refactor and v-tab adding to details-tile
